op-batcher: Reorganize files
This places most of the files into a single folder.
