+++ op-batcher/batcher/channel_manager.go
@@ -1,4 +1,4 @@
-package op_batcher
+package batcher
 
 import (
 	"bytes"
