+++ op-batcher/batcher/driver.go
@@ -1,4 +1,4 @@
-package op_batcher
+package batcher
 
 import (
 	"context"
@@ -12,7 +12,6 @@ import (
 	"time"
 
 	hdwallet "github.com/ethereum-optimism/go-ethereum-hdwallet"
-	"github.com/ethereum-optimism/optimism/op-batcher/sequencer"
 	"github.com/ethereum-optimism/optimism/op-node/eth"
 	"github.com/ethereum-optimism/optimism/op-proposer/txmgr"
 	"github.com/ethereum/go-ethereum/accounts"
@@ -27,7 +26,7 @@ import (
 type BatchSubmitter struct {
 	txMgr *TransactionManager
 	addr  common.Address
-	cfg   sequencer.Config
+	cfg   DriverConfig
 	wg    sync.WaitGroup
 	done  chan struct{}
 	log   log.Logger
@@ -141,7 +140,7 @@ func NewBatchSubmitterWithSigner(cfg Config, addr common.Address, signer SignerF
 		SafeAbortNonceTooLowCount: cfg.SafeAbortNonceTooLowCount,
 	}
 
-	batcherCfg := sequencer.Config{
+	batcherCfg := DriverConfig{
 		Log:               l,
 		Name:              "Batch Submitter",
 		L1Client:          l1Client,
