+++ op-batcher/cmd/main.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/urfave/cli"
 
-	batcher "github.com/ethereum-optimism/optimism/op-batcher"
+	"github.com/ethereum-optimism/optimism/op-batcher/batcher"
 	"github.com/ethereum-optimism/optimism/op-batcher/flags"
 	oplog "github.com/ethereum-optimism/optimism/op-service/log"
 	"github.com/ethereum/go-ethereum/log"
